Remove redundant cast in MultiModalEdge ctor.

+++ reader-gtfs/src/main/java/com/graphhopper/gtfs/GraphExplorer.java
@@ -187,9 +187,13 @@ public final class GraphExplorer {
                                         + ", using default weight.");
                             }
                         }
-                        action.accept(new MultiModalEdge(e.getEdge(), e.getBaseNode(), e.getAdjNode(),
-                                (long) (connectingWeighting.calcEdgeMillis(e.detach(false), reverse)), weight,
-                                e.getDistance()));
+                        action.accept(
+                                new MultiModalEdge(e.getEdge(), e.getBaseNode(),
+                                        e.getAdjNode(),
+                                        connectingWeighting.calcEdgeMillis(
+                                                e.detach(false), reverse),
+                                        weight,
+                                        e.getDistance()));
                         return true;
                     }
                 }
